# src/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "types.DependencyGraph": {
                "properties": {
                    "nodes": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/types.DependencyNode"
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "types.DependencyNode": {
                "properties": {
                    "depends_on": {
                        "type": "object"
                    },
                    "is_ready": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "process_status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "info": {
        "contact": {
            "name": "Process Compose Discord Channel",
            "url": "https://discord.gg/S4xgmRSHdC"
        },
        "description": "This is a sample Process Compose server.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "Process Compose API",
        "version": "1.75.2"
    },
    "openapi": "3.0.3",
    "paths": {
        "/graph": {
            "get": {
                "description": "Returns the process dependency graph with current status",
                "operationId": "GetDependencyGraph",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.DependencyGraph"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Get dependency graph",
                "tags": [
                    "Graph"
                ]
            }
        },
...
        "/namespace": {
            "delete": {
                "description": "Delete all processes from current config in the given namespace",
                "operationId": "DeleteNamespace",
                "parameters": [
                    {
                        "description": "Namespace Name",
                        "in": "query",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "All processes removed, may be zero if non existent"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Some processes failed to be removed, can happen if some have dependants or removed concurrently"
                    }
                },
                "summary": "Delete namespace processes",
                "tags": [
                    "Namespace"
                ]
            }
        },
        "/namespace/disable/{name}": {
            "patch": {
                "description": "Disables all processes in the given namespace",
                "operationId": "DisableNamespace",
                "parameters": [
                    {
                        "description": "Namespace Name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "All processes in namespace disabled"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Some processes in namespace failed to be disabled, can happen if several opposite updates to same namespace are happening"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "No processes in namespace"
                    }
                },
                "summary": "Disable all processes in a namespace",
                "tags": [
                    "Namespace"
                ]
            }
        },
        "/namespace/enable/{name}": {
            "patch": {
                "description": "Enables all processes in the given namespace",
                "operationId": "EnableNamespace",
                "parameters": [
                    {
                        "description": "Namespace Name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "All processes in namespace enabled"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Some processes in namespace failed to be enabled"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "No processes in namespace"
                    }
                },
                "summary": "Enable all processes in a namespace",
                "tags": [
                    "Namespace"
                ]
            }
        },
        "/namespace/stop/{name}": {
            "patch": {
                "description": "Sends kill signal to all processes in the given namespace",
                "operationId": "StopNamespace",
                "parameters": [
                    {
                        "description": "Namespace Name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Stopped All Processes in Namespace"
                    },
                    "207": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Stopped Part of Processes in Namespace"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Failed to stop some processes, they may have some dependants"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "No proccesses in namespace"
                    }
                },
                "summary": "Stop all processes in a namespace",
                "tags": [
                    "Namespace"
                ]
            }
        },
        "/process": {
            "post": {
                "description": "Update process",
                "operationId": "UpdateProcess",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/types.ProcessConfig"
                            }
                        }
                    },
                    "description": "Process configuration to update",
                    "required": true,
                    "x-originalParamName": "process"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.ProcessConfig"
                                }
                            }
                        },
                        "description": "Updated Process Config"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Updates process configuration",
                "tags": [
                    "Process"
                ]
            }
        },
        "/process/info/{name}": {
            "get": {
                "description": "Retrieves the given process and its config",
                "operationId": "GetProcessInfo",
                "parameters": [
                    {
                        "description": "Process Name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.ProcessConfig"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Get process config",
                "tags": [
                    "Process"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "https://localhost:8080/"
        }
    ]
}